LayoutTests/imported/w3c/web-platform-tests/css/css-ui/parsing/outline-color-computed-expected.txt


PASS Property outline-color value 'currentColor'
PASS Property outline-color value 'red'
FAIL auto computes to itself or to currentColor depending on the value of outline-style assert_equals: expected "auto" but got "rgb(255, 0, 0)"
